+++ lib/recording_studio_support.rb
@@ -1,7 +1,11 @@
 # frozen_string_literal: true
 
+require "active_record"
 require "recording_studio"
 require "recording_studio_accessible"
+require "recording_studio_attachable"
+require "recording_studio_trashable"
+require "recording_studio_orderable"
 require "recording_studio_support/version"
 require "recording_studio_support/engine"
 require "recording_studio_support/configuration"